position

+++ tools.h
@@ -5,7 +5,7 @@
 
 wxString GetMsg(int id);
 void SetLangId(int id);
-wxString ConvertDegree(float degree);
+wxString ConvertDegree(double degree);
 wxString FormatLongitude(float x); 
 wxString FormatLatitude(float y);
 double nvDistance(double lon1, double lat1, double lon2, double lat2, int distanceunit);
